Hash file checker and agreements

+++ datamaestro/utils.py
@@ -4,7 +4,9 @@ import os.path as op
 import json
 from pathlib import PosixPath, Path
 from shutil import rmtree
+import shutil
 import hashlib
+from tqdm import tqdm
 
 class TemporaryDirectory:
     def __init__(self, path: Path):
@@ -20,41 +22,107 @@ class TemporaryDirectory:
         if self.delete:
             rmtree(self.path)
 
-class HashCheck:
+class FileChecker():
+    def check(self, path: Path):
+        """Check if the file is correct and throws an exception if not"""
+        raise NotImplementedError()
+
+class HashCheck(FileChecker):
     """Check a file against a hash"""
     def __init__(self, hashstr: str, hasher=hashlib.md5):
         self.hashstr = hashstr
         self.hasher = hasher
 
-    def check(self, path: Path) -> bool:
+    def check(self, path: Path):
         """Check the given file
 
         returns true if OK
         """        
-        with path.open() as fp:
+        with path.open("rb") as fp:
             hasher = self.hasher()
-            hasher.update(fp)
+            chunk = fp.read(8192)
+            while chunk:
+                hasher.update(chunk)
+                chunk = fp.read(8192)
         s = hasher.hexdigest()
         if s != self.hashstr:
-            raise Exception(f"Digest do not match ({self.hashcheck} vs {s})")
+            raise IOError(f"Digest do not match ({self.hashstr} vs {s})")
 
 
 class CachedFile:
-    """Represents a downloaded file that has been cached"""
+    """Represents a downloaded file that has been cached
+    
+    The file is automatically deleted when closed if keep is False
+    and used is True
+    """
 
-    def __init__(self, path, keep=False):
+    def __init__(self, path, keep=False, others=[]):
         self.path = path
         self.keep = keep
+        self._force_delete = False
+        self.others = []
 
     def __enter__(self):
         return self
 
+    def force_delete(self):
+        """Force the file to be deleted (even if an exception was thrown)"""
+        self.force_delete = True
+
     def __exit__(self, exc_type, exc_value, traceback):
+        # Avoid removing the file if an exception was thrown
+        if not self.force_delete and exc_type is not None:
+            logging.info("Keeping cache file %s (exception thrown)", self.path)
+            return
+
         try:
             if not self.keep:
+                logging.info("Deleting cache file %s", self.path)
                 self.path.unlink()
+                for other in self.others:
+                    other.unlink()
         except Exception as e:
             logging.warning("Could not delete cached file %s [%s]", self.path, e)
+
+class CurlDownloadReportHook(tqdm):
+    """Report hook for tqdm when downloading from the Web with PyCURL"""
+
+    def __init__(self, **kwargs):
+        kwargs.setdefault("unit", "B")
+        kwargs.setdefault("unit_scale", True)
+        kwargs.setdefault("miniters", 1)
+        super().__init__(**kwargs)
+
+    def __call__(self, download_total, downloaded, upload_total, uploaded):
+        """PyCURL callback"""
+        if download_total is not None:
+            self.total = download_total
+        self.update(downloaded - self.n)  # will also set self.n = b * bsize
+
+def downloadURL(url: str, path: Path, resume: bool=False):
+    import pycurl
+    try:
+        with path.open(f"ab") as fp, CurlDownloadReportHook(desc="Downloading %s" % url) as reporthook:
+            c = pycurl.Curl()
+            c.setopt(pycurl.URL, url)
+            if resume:
+                c.setopt(pycurl.RESUME_FROM, os.path.getsize(path))
+            c.setopt(pycurl.WRITEDATA, fp)
+            c.setopt(pycurl.NOPROGRESS, 0)
+            c.setopt(pycurl.FOLLOWLOCATION, 1)
+            c.setopt(pycurl.MAXREDIRS, 5)
+            c.setopt(pycurl.XFERINFOFUNCTION, reporthook)
+            c.perform()
+            fp.close()
+            shutil.move(path, dlpath)
+    except pycurl.error as e:
+        code = e.args[0]
+        message = e.args[1]
+        if code == pycurl.E_RANGE_ERROR:
+            logging.error("Cannot resume download (%s) - starting all over", message)
+            path.unlink()
+            downloadURL(url, path, False)
+        raise
 
 
 def deprecated(message, f):
